+++ smartcontract/cli/src/tenant/mod.rs
@@ -5,3 +5,4 @@ pub mod get;
 pub mod list;
 pub mod remove_administrator;
 pub mod update;
+pub mod update_payment_status;