+++ smartcontract/programs/doublezero-serviceability/src/processors/tenant/mod.rs
@@ -3,3 +3,4 @@ pub mod create;
 pub mod delete;
 pub mod remove_administrator;
 pub mod update;
+pub mod update_payment_status;