+++ smartcontract/sdk/rs/src/commands/tenant/update_payment_status.rs
@@ -1,25 +1,23 @@
 use crate::DoubleZeroClient;
 use doublezero_serviceability::{
     instructions::DoubleZeroInstruction, pda::get_globalstate_pda,
-    processors::tenant::update::TenantUpdateArgs,
+    processors::tenant::update_payment_status::UpdatePaymentStatusArgs,
 };
 use solana_sdk::{instruction::AccountMeta, pubkey::Pubkey, signature::Signature};
 
 #[derive(Debug, PartialEq, Clone)]
-pub struct UpdateTenantCommand {
+pub struct UpdatePaymentStatusCommand {
     pub tenant_pubkey: Pubkey,
-    pub vrf_id: Option<u16>,
-    pub token_account: Option<Pubkey>,
+    pub payment_status: u8,
 }
 
-impl UpdateTenantCommand {
+impl UpdatePaymentStatusCommand {
     pub fn execute(&self, client: &dyn DoubleZeroClient) -> eyre::Result<Signature> {
         let (globalstate_pubkey, _) = get_globalstate_pda(&client.get_program_id());
 
         client.execute_transaction(
-            DoubleZeroInstruction::UpdateTenant(TenantUpdateArgs {
-                vrf_id: self.vrf_id,
-                token_account: self.token_account,
+            DoubleZeroInstruction::UpdatePaymentStatus(UpdatePaymentStatusArgs {
+                payment_status: self.payment_status,
             }),
             vec![
                 AccountMeta::new(self.tenant_pubkey, false),